services/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
requests==2.31.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6

